[BUGFIX] Fix SoapFault response
Using $server->fault will cause a http 500 status code to be sent which in turn
will trigger the web server's error page to be shown, instead of the SOAP XML
The header sent above will just be ignored.
Because of that, we forge a SOAP Fault XML ourselves and just echo it.

flush() to prevent SoapServer to change the status code obviously also does not work
This fails when some other kind of output buffering is in place (e.g. for gzip compression)
in the web server.


git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/ter/trunk@79807 735d13b6-9817-0410-8766-e36946ffe9aa