[BUGFIX] Correct dependencies if only one is transmitted
If an extension has only one dependency and all is transmitted
thorugh a SOAP request, the dependencies property is an object and not
an array as expected.

All following code will then act as no dependency was set.
To fix that, we need to make the property an array again.

Change-Id: I0f89013a3b17699634ebda845b072e6d4d86f79e
Reviewed-on: https://review.typo3.org/43046
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>